backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1
openai>=2.0.0
tiktoken==0.8.0
openpyxl==3.1.5
fpdf2==2.8.1
